refactor and split up code
