+++ src/assets/blockNavigation.css
@@ -23,149 +23,3 @@ div[role="button"]:not([aria-expanded]):not([aria-pressed]):active {
 li[role="presentation"]:active {
     pointer-events: none !important;
 }
-
-/* styling link preview popup */
-.pdftron-link-popup-outer {
-    display: none;
-    position: absolute;
-    z-index: 1000000;
-    padding: 10px 0;
-    background-color: transparent;
-    cursor: default;
-}
-
-.pdftron-link-popup-inner {
-    font-family: inherit;
-    font-size: 12px;
-    text-transform: initial;
-    font-weight: 400;
-    line-height: 14.4px;
-    /* background is white so color cannot be inherit (parent color might be white) */
-    /* important to overide any hover behaviour from parent */
-    color: #000 !important;
-    background-color: white !important;
-    border-radius: 4px;
-    box-shadow: 0px 0px 3px 0px #868E96;
-    /* box-shadow: 0 1px 3px 1px rgba(60, 64, 67, 0.15); */
-    padding: 12px;
-    text-align: left;
-    width: 260px;
-    overflow-wrap: break-word;
-    white-space: normal;
-}
-
-.pdftron-link-popup-inner :hover {
-    background-color: initial;
-}
-
-.pdftron-link-popup-inner .pdftron-link-favicon,
-.pdftron-link-popup-inner .pdftron-link-favicon-empty {
-    margin-inline-end: 4px;
-    width: 20px;
-    max-height: 20px;
-    flex: 0 0 20px;
-}
-
-.pdftron-link-popup-inner .pdftron-link-favicon {
-    display: none;
-}
-
-.pdftron-link-popup-inner .pdftron-link-favicon-empty {
-    height: 20px;
-    border-radius: 4px;
-    background-color: #E0E0E0;
-}
-
-.pdftron-link-popup-inner .pdftron-link-title {
-    font-weight: 600;
-    color: #00A5E4 !important;
-    flex-grow: 1;
-    flex-shrink: 1;
-    text-overflow: ellipsis;
-    overflow: hidden;
-    white-space: nowrap;
-    margin-inline-end: 11px;
-    cursor: pointer;
-}
-
-.pdftron-link-popup-inner .pdftron-link-title:hover,
-.pdftron-link-popup-inner .pdftron-link-title:active {
-    color: #1A4971 !important;
-    text-decoration: none !important;
-}
-
-.pdftron-link-popup-inner .pdftron-link-button-copy {
-    color: #868E96;
-    background-color: transparent;
-    padding: 3px;
-    margin: 0;
-    border: 0;
-    cursor: pointer;
-    position: relative;
-    flex: 0 0 20px;
-    outline: none;
-}
-
-.pdftron-link-popup-inner .pdftron-link-button-copy svg {
-    /* override https://www.roadandtrack.com/car-culture/a38378639/mate-rimac-profile/ */
-    width: auto !important;
-    height: auto !important;
-    vertical-align: middle !important;
-}
-
-.pdftron-link-popup-inner .pdftron-link-button-copy:hover {
-    background-color: #E7EDF3;
-}
-
-.pdftron-link-popup-inner .pdftron-link-button-copy:active {
-    color: #1A4971;
-    background-color: #DDE6EE;
-}
-
-.pdftron-link-popup-inner .pdftron-link-button-copy::before,
-.pdftron-link-popup-inner .pdftron-link-button-copy::after {
-    display: block;
-    opacity: 0;
-    visibility: hidden;
-    position: absolute;
-}
-
-.pdftron-link-popup-inner .pdftron-link-button-copy::before {
-    content: attr(data-title);
-    top: 26px;
-    left: -31px;
-    padding: 8px 12px;
-    color: #fff;
-    background-color: #222;
-    border-radius: 4px;
-    white-space: nowrap;
-}
-
-.pdftron-link-popup-inner .pdftron-link-button-copy::after {
-    content: "";
-    width: 0;
-    height: 0;
-    top: 21px;
-    left: 50%;
-    margin-left: -8px;
-    border-left: 8px solid transparent;
-    border-right: 8px solid transparent;
-    border-bottom: 6px solid #222;
-}
-
-.pdftron-link-popup-inner .pdftron-link-button-copy:hover::before,
-.pdftron-link-popup-inner .pdftron-link-button-copy:hover::after {
-    visibility: visible;
-    opacity: 1;
-    transition: opacity .3s 1s ease-out;
-}
-
-.pdftron-link-popup-inner .pdftron-link-meta {
-    margin-top: 8px;
-    display: -webkit-box;
-    -webkit-line-clamp: 2;
-    line-clamp: 2;
-    -webkit-box-orient: vertical;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
