+++ src/assets/getTextData.js
@@ -1,3 +1,4 @@
+/* eslint-disable no-undef */
 const getTextData = (body) => {
   const traverseTextNode = (parentNode, struct, offsets, quads, str, linksArray) => {
     const range = document.createRange();
@@ -143,28 +144,6 @@ const isInvalidNode = (node) => {
   return (!node) || (node.getBoundingClientRect && (node.getBoundingClientRect().width === 0 || node.getBoundingClientRect().height === 0));
 };
 
-const getPageHeight = () => {
-  let sum = 0;
-  // for some web pages, <html> and <body> have height: 100%
-  // sum up the <body> children's height for an accurate page height
-  // example: when page expands and then shrinks, <body> height will not reflect this change since it's 100% (which is the iframe height)
-  document.body.childNodes.forEach((el) => {
-    if (el.nodeType === Node.ELEMENT_NODE) {
-      const style = window.getComputedStyle(el);
-      // filter hidden/collapsible elements
-      if (style.display === 'none' || style.visibility === 'hidden' || style.opacity === '0' || style.position === 'fixed' || style.position === 'absolute') {
-        return;
-      }
-      // some elements have undefined clientHeight
-      // favor scrollHeight since clientHeight does not include padding
-      if (!isNaN(el.scrollHeight) && !isNaN(el.clientHeight)) {
-        sum += (el.clientHeight > 0 ? (el.scrollHeight || el.clientHeight) : el.clientHeight);
-      }
-    }
-  });
-  return sum;
-};
-
 const getClientUrl = () => {
   const { origin } = new URL(document.referrer);
   return origin;
@@ -193,12 +172,7 @@ document.addEventListener('DOMContentLoaded', () => {
   const observer = new MutationObserver(() => {
     debounceSendDataOnMutation();
   });
-  observer.observe(document.body, {
-    attributes: true,
-    childList: true,
-    subtree: true,
-    characterData: true,
-  });
+  observer.observe(document.body, mutationObserverConfig);
 });
 
 document.addEventListener('transitionend', () => {
