+++ types/utils/getHostPortSSL.d.ts
@@ -1,3 +1,3 @@
-import type { ServerHostPortSSL } from './data';
+import type { ServerHostPortSSL } from './types';
 declare const getHostPortSSL: (url: string, allowHTTPProxy?: boolean) => ServerHostPortSSL;
 export { getHostPortSSL };
